Stepless riding zoom and earlier, speed-stretched announcements
Two ride reports from the same tour.

The riding camera sat too close to see what was coming: the whole point
of a curve navigator is reading the next bend, and at touring speed the
map showed barely more than the junction underneath. CameraController
now interpolates a fractional zoom between speed anchors instead of
snapping to whole levels with a hysteresis band, and the anchors moved
out by roughly one level throughout. Being stepless also removes the
band edges, which were the other thing that read as a jump; the eased
per-fix approach replaces the hysteresis that existed to hide them.
Zoom is a Double from the controller through to the MapLibre camera,
which took fractional zoom all along.

Announcements arrived late. Two causes the plain "N seconds before the
manoeuvre" model ignores: the words themselves take time (a helmet
headset needs up to 1.5 s to open its audio path, then the sentence
runs 2-3 s), and a faster rider needs disproportionately more warning,
not proportionally more. AnnouncementTiming now pays for the speech on
every tier and stretches each tier's lead with speed - fully on the
early heads-up, barely at all on the last-chance call, which has to
stay near the junction to mean "now". At 100 km/h the tiers land at
roughly 640/260/115 m instead of 420/196/84, which is the Landstrasse
cadence the brief asks for; per-tier distance floors keep the calls
useful in stop-and-go traffic too.

+++ app/src/main/java/com/motoroute/ui/OpenCurvRoot.kt
@@ -270,7 +270,7 @@ private fun MapRoot(
     manualZoom: Boolean,
     demoRunning: Boolean,
     speedCameraWarning: com.motoroute.domain.cameras.SpeedCameraWarning?,
-    zoom: Int,
+    zoom: Double,
     onOpenData: () -> Unit,
     onOpenSettings: () -> Unit,
     onOpenSearch: (SearchMode) -> Unit,
